requirements.txt
# 阿里 LangStudio Agent 示例依赖
# Python >= 3.9

fastapi>=0.100.0
uvicorn[standard]>=0.22.0
openai>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
ag-ui-protocol>=0.1.0
python-dotenv>=1.0.0
requests>=2.28.0
